+++ Assets/Scripts/Testing/HitArea.cs
@@ -7,11 +7,11 @@ namespace Testing
     public class HitArea : MonoBehaviour, IHittable
     {
         [MinValue(0)]
-        [SerializeField] private int _damage;
+        [SerializeField] private float _damage;
         [MinValue(0)]
         [SerializeField] private float _interval;
 
-        public int Damage => _damage;
+        public float Damage => _damage;
         public float Interval => _interval;
     }
 }
